Added the ability to download files.
This is not a great system so far. You have to select and download the
files one at a time. I'd like to be able to select multiple files at once
and then have it be downloaded in a single zipped file. The primary reason
why I'd want to download files is to send it to other people. so only so
helpful as is.