Fixed Threading and tempdir.
Threading now works properly. The log file, once it starts loading, will
load endlessly, this means it will even pickup new changes to the file
that have happened since the file was initially been created. That is
assuming the changes are addendums, reasonable assumption for a log file.

Fixed the system to properly use a temporary directory. No more hard coded
directory that neither python nor the OS want to dispose of. Hurray!