Dream feature incoming! Logback editing!
So I decided I wanted to do it. It will be such a powerful feature that I
just had to do it. Logback.py has been edited to be able to parse a
logback.xml file, turn it into objects, and then turn around and spit it
out as a new logback file. This is not quite completing of the logback
editing that I envision, but it is certainly a start.

Honestly, to get this to work I'm kinda going to need to re-create a lot
of the interface. I know that's coming really what ever I do. I've painted
myself into a corner. But as my first time making a GUI, I'm not to upset.