Well I am super pleased!
Today has been very good. I got a lot done on Octo and it will be quite a
nice tool. I have decided to set it up in a way that will make it much
easier to write new features for it. Last time I got locked into the
format I had choosen for viewing logs and it hindered any growth. New
features just didn't have a place. This time I've stepped back and now
look at logging as one of many features.

The idea is that you have a tab for each connection to a server, and you
can have multiple connections to a server. Then you have a page for each
tool you want to use on that server. Right now I've got "Info" as the only
tool, which doesn't help me view logs, but! it is a start. Next time I'll
make a log viewer.

Well, actually next time I'll probably also make the logback adjuster.
That one is super helpful for getting the info you need.

Other thoughts:

A panel that talks about each of the loaded portlets and options to reload
one from a specific branch of the code base.

A more indepth panel that just shows tomcat stuff.

Just catalina.out. That would actually be pretty useful to just have
catalina.out right there and handy.

And I'm sure I'll come up with others as I use it, but the nice thing is
that it will be super easy to do now!