Added Catalina out panel.
Also, got a little more done with the logback panel. have a ways to go to
get it all properly hooked up, but it's starting to be a little more fully
fledged. Right now I plan on spending the rest of my day trying to get it
to compile for android! that should be fun. O_O