Logback now edits!
Got logback up to an alpha status. It can do basic edits, i.e. turn
loggers on and off, but nothing to complex yet. Also added a few new
features to the connection class. Made it a little more powerful to keep
the rest of the program nice and simple.

Comments! lots of them. I'm trying to keep up on the documentation as I
go.