Fix for python version and on_pause
I found a little article talking about on pause. I added in a dummy class
that should remedy the sleeping issue. Basically when on_pause is called
you need to say "Yes, save me!" and all will be good. There is a
reciprical method, on_resume, that you can call to restore any data that
is of particular interest to be restored.

Also, fixed the multiple python versions issue.