Got a settings window.
I'm still working on making it everything I want it to be. I'm having to
make some custom extensions of the settings classes to get it to work, but
I think I'll be able to hide the adding and removing of servers away in
the settings panel. It will give this project a real nice polished feel.

Right now I've stopped on the advent of making a pop up. I'll get back to
this when I've got the chance.