Settings a work!
I've got setting working now. You can add servers on the fly. I also got a
basic password salting system working. right now it's not very good as it
just uses "63" as the salt. I need to find a way of getting a value that
is unique, maybe a password from the user, to find a way to properly salt
the passwords. But the basics I've got here will work for now. Passwords
are at least obfescated.