SQL panel is working!
The last parts of it just fell neatly in to place. Some times when I'm
working on the ender lying components I can feel like I've lost myself in
the forest. So much so that I have no idea how much longer until a
features is finished. When all of a sudden, BAM! feature is full
functional. It is a great feeling in the end, but so discouraging in the
middle.

This panel may need some extra work, but so far kivy has managed all
threading issues I could be concerned about. At least as goes performance.
I keep in mind to put mutex and locks in places to make sure that nothing
un-expected happens. But I'm happy to not have to worry about figuring out
how to thread. Octo is almost back to a useable status.