Fixed some catalina and logback problems.
The catalina panel would stop displaying content after a while. This was,
i think, due to some sort of over flow for in the label. I fixed this by
making each line its own label. I also got it to dynamically reflow the
text if it had been changed. The process could use some improving I think
because I have to iterate through every single line and update it's width,
but it is progress never the less.

I also fixed a problem with the logback interface. appearently my simple
sudo command needs to have a readline called for it to work properly. Now
the toggling of logback files works as it should.

One thing worth noting: The logback panel and the sql panel both reference
the logback file, but do so seperately. It would be nice to join them so
they can keep each other up to date.