Added an icon!
My first time trying to do an icon. I just made something quick and simple
with word art. Updated the readme a bit. Moved the features that were
lost when I made the move to Kiv back to the "To work on" side. Also
updated the spec to say this is version 0.3.